fix: changed Wall class into TS
