--- srcs/games/pong/backend/plugins/game/classes/Wall.js
@@ -1,32 +1,0 @@
-
-import { Vector3 } from "@babylonjs/core";
-import Paddle from "./Paddle.js";
-
-export default class Wall extends Paddle {
-    constructor(scene, name, position = new Vector3(0, 0, 0), options = {}) {
-        super(scene, name, position, options);
-        this._keys = {};
-    }
-
-    _setupInput() {
-        // Do nothing
-    }
-
-    getPassThrough() {
-        return this._passThrough;
-    }
-
-    calculateBounce(ball)	{
-        if (!this._passThrough) {
-            if (ball.getLastHit() === this.name) return ;
-            ball.setLastHit(this.mesh.name)
-            ball._direction.z *= -1;
-            ball._direction.normalize();
-        }
-        else {
-            ball.mesh.position.set(0, 0, 0);
-            ball.setSpeed(0);
-            ball.setLastHit("");
-        }
-    }
-}
